backend/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.2
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0